+++ examples/poem/cookie-session/src/main.rs
@@ -6,8 +6,9 @@
 use poem::{
     get, handler,
     listener::TcpListener,
+    middleware::CookieJarManager,
     web::cookie::{Cookie, CookieJar},
-    Route, Server,
+    EndpointExt, Route, Server,
 };
 
 #[handler]
@@ -34,7 +35,9 @@ async fn main() -> Result<(), std::io::Error> {
     }
     tracing_subscriber::fmt::init();
 
-    let app = Route::new().at("/", get(count));
+    let app = Route::new()
+        .at("/", get(count))
+        .with(CookieJarManager::new());
     let listener = TcpListener::bind("127.0.0.1:3000");
     let server = Server::new(listener).await?;
     server.run(app).await