+++ poem/src/lib.rs
@@ -75,5 +75,5 @@ pub use poem_derive::handler;
 pub use request::{Request, RequestBuilder, RequestParts};
 pub use response::{Response, ResponseBuilder, ResponseParts};
 pub use route::{connect, delete, get, head, options, patch, post, put, trace, Route, RouteMethod};
-pub use server::{warps_endpoint, Server};
+pub use server::Server;
 pub use web::{FromRequest, IntoResponse, RequestBody};