+++ poem/src/middleware/cookie_jar_manager.rs
@@ -1,12 +1,31 @@
-use crate::{web::cookie::CookieJar, Endpoint, IntoResponse, Middleware, Request, Response};
+use std::sync::Arc;
+
+use crate::{
+    web::cookie::{CookieJar, CookieKey},
+    Endpoint, IntoResponse, Middleware, Request, Response,
+};
 
 /// Middleware for CookieJar support.
-///
-/// NOTE: You usually don't need to add this middleware yourself, because
-/// `poem::Server` will automatically add it when the `cookie` feature is
-/// enabled.
 #[derive(Default)]
-pub struct CookieJarManager;
+pub struct CookieJarManager {
+    key: Option<Arc<CookieKey>>,
+}
+
+impl CookieJarManager {
+    /// Creates a new `Compression` middleware.
+    #[must_use]
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Specify the `CookieKey` used for the `CookieJar::private` and
+    /// `CookieJar::signed` methods.
+    pub fn with_key(key: CookieKey) -> Self {
+        Self {
+            key: Some(Arc::new(key)),
+        }
+    }
+}
 
 impl<E> Middleware<E> for CookieJarManager
 where
@@ -15,13 +34,17 @@ where
     type Output = CookieJarManagerEndpoint<E>;
 
     fn transform(&self, ep: E) -> Self::Output {
-        CookieJarManagerEndpoint { inner: ep }
+        CookieJarManagerEndpoint {
+            inner: ep,
+            key: self.key.clone(),
+        }
     }
 }
 
 /// Endpoint for CookieJarManager middleware.
 pub struct CookieJarManagerEndpoint<E> {
     inner: E,
+    key: Option<Arc<CookieKey>>,
 }
 
 #[async_trait::async_trait]
@@ -29,7 +52,9 @@ impl<E: Endpoint> Endpoint for CookieJarManagerEndpoint<E> {
     type Output = Response;
 
     async fn call(&self, mut req: Request) -> Self::Output {
-        let cookie_jar = CookieJar::extract_from_headers(req.headers());
+        let mut cookie_jar = CookieJar::extract_from_headers(req.headers());
+        cookie_jar.key = self.key.clone();
+
         if req.state().cookie_jar.is_none() {
             req.state_mut().cookie_jar = Some(cookie_jar.clone());
             let mut resp = self.inner.call(req).await.into_response();
@@ -44,7 +69,7 @@ impl<E: Endpoint> Endpoint for CookieJarManagerEndpoint<E> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{handler, http::StatusCode, EndpointExt};
+    use crate::{handler, http::StatusCode, web::cookie::Cookie, EndpointExt};
 
     #[tokio::test]
     async fn test_cookie_jar_manager() {
@@ -53,9 +78,47 @@ mod tests {
             assert_eq!(cookie_jar.get("value").unwrap().value_str(), "88");
         }
 
-        let ep = index.with(CookieJarManager);
+        let ep = index.with(CookieJarManager::new());
         let resp = ep
             .call(Request::builder().header("Cookie", "value=88").finish())
+            .await;
+        assert_eq!(resp.status(), StatusCode::OK);
+    }
+
+    #[tokio::test]
+    async fn test_cookie_jar_manager_with_key() {
+        #[handler(internal)]
+        async fn index(cookie_jar: &CookieJar) {
+            assert_eq!(
+                cookie_jar.private().get("value1").unwrap().value_str(),
+                "88"
+            );
+            assert_eq!(cookie_jar.signed().get("value2").unwrap().value_str(), "99");
+        }
+
+        let key = CookieKey::generate();
+
+        let ep = index.with(CookieJarManager::with_key(key.clone()));
+        let cookie_jar = CookieJar::default();
+        cookie_jar
+            .private_with_key(&key)
+            .add(Cookie::new_with_str("value1", "88"));
+        cookie_jar
+            .signed_with_key(&key)
+            .add(Cookie::new_with_str("value2", "99"));
+        let resp = ep
+            .call(
+                Request::builder()
+                    .header(
+                        "Cookie",
+                        format!(
+                            "{}; {}",
+                            cookie_jar.get("value1").unwrap().to_string(),
+                            cookie_jar.get("value2").unwrap().to_string()
+                        ),
+                    )
+                    .finish(),
+            )
             .await;
         assert_eq!(resp.status(), StatusCode::OK);
     }