+++ poem/src/server.rs
@@ -16,8 +16,6 @@ use tokio::{
 };
 use tracing::{Instrument, Level};
 
-#[cfg(feature = "cookie")]
-use crate::middleware::CookieJarManager;
 use crate::{
     listener::{Acceptor, Listener},
     web::RemoteAddr,
@@ -60,7 +58,7 @@ impl<T: Acceptor> Server<T> {
         signal: impl Future<Output = ()>,
         timeout: Option<Duration>,
     ) -> IoResult<()> {
-        let ep = warps_endpoint(ep.into_endpoint());
+        let ep = ep.into_endpoint();
         let ep = Arc::new(ep.map_to_response());
         let Server { mut acceptor } = self;
         let alive_connections = Arc::new(AtomicUsize::new(0));
@@ -180,11 +178,4 @@ async fn serve_connection(
     #[cfg(feature = "websocket")]
     let conn = conn.with_upgrades();
     let _ = conn.await;
-}
-
-#[doc(hidden)]
-pub fn warps_endpoint(ep: impl Endpoint) -> impl Endpoint<Output = Response> {
-    #[cfg(feature = "cookie")]
-    let ep = ep.with(CookieJarManager);
-    ep.map_to_response()
 }